requirements.txt
asgiref==3.2.3
distro==1.4.0
Django==3.0.3
mysqlclient==1.4.6
psutil==5.7.0
py-cpuinfo==5.0.0
python-nmap==0.6.1
pytz==2019.3
sqlparse==0.3.0
